Added batch functionality
Added many rewrites to how the program handles making and running args. Now the old usecompiler function is used to get args for a single file which are put into an arraylist which is passed to the usecmd method. The usecmd function now constructs string with all the arguments with && between them so that they are ran in sequence.
Also includes the disam fix.